requirements.txt
pytorch-lightning>=1.5
torch>=1.10
torchvision
opencv-python
mediapipe
scikit-learn
scikit-image
pyyaml
matplotlib
numpy 